rename file chat -> chatTypes and update imports
